+++ app/config.py
@@ -12,6 +12,7 @@ class TestingConfig(Config):
     TESTING = True
 
 class ProductionConfig(Config):
+    pass
     
 configuration = {
     "development" : DevelopmentConfig,
